\name{EBSeq_NingLeng-package}
\alias{EBSeq_NingLeng-package}
\alias{EBSeq_NingLeng}
\docType{package}
\title{
EBSeq: RNA-Seq Differential Expression Analysis on both gene and isoform level 
}
\description{
A Negative Binomial - beta model was built to analyze the RNASeq data.
We used the empirical bayes method and EM algrithom.
}
\details{
\tabular{ll}{
Package: \tab EBSeq_NingLeng\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-13\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ning Leng

Maintainer: Ning Leng  <nleng@wisc.edu>
}
\references{
}
\keyword{ package }
\seealso{

}
\examples{
}
