\name{MultiGeneMat}
\alias{MultiGeneMat}
\docType{data}
\title{
The simulated data for multiple condition gene DE analysis
}
\description{
'MultiGeneMat' generates a set of the simulated data for multiple condition gene DE analysis.
}
\usage{data(MultiGeneMat)}
\source{
Ning Leng, John A. Dawson, James A. Thomson, Victor Ruotti, Anna I. Rissman, Bart M.G. Smits, Jill D. Haag, Michael N. Gould, Ron M. Stewart, and Christina Kendziorski. EBSeq: An empirical Bayes hierarchical model for inference in RNA-seq experiments. Bioinformatics (2013)
}
\seealso{
	GeneMat
}
\examples{
data(MultiGeneMat)
}
\keyword{datasets}
